HotelFeatures +

+++ src/main/java/com/tobeto/hotel_booking_java4a_pair5/controllers/HotelsController.java
@@ -209,7 +209,7 @@ public class HotelsController {
                                                                                        @RequestParam(required = false) Integer roomCapacity,
                                                                                               @RequestParam(required = false) Double minPrice,
                                                                                               @RequestParam(required = false) Double maxPrice,
-                                                                                              @RequestParam(required = false) List<Long> featureIds) {
+                                                                                              @RequestParam(required = false) List<Integer> featureIds) {
         List<Hotel> hotels = hotelService.searchAllHotelsWithFilters(location, startDate, endDate, roomCapacity,minPrice,maxPrice,featureIds);
         List<FindHotelWithAvailableRoomsResponse> response = HotelMapper.INSTANCE.findHotelResponseFromHotelResponseList(hotels);
